+++ simulation/include/SimCalcFuncs.cuh
@@ -14,7 +14,7 @@ namespace RunCudaFunctions
 {
     void run_calculate_streaming(dim3 blocks, dim3 threads, LatticeData lattice, LatticeData templattice);
 
-    void run_calculate_collision(dim3 blocks, dim3 threads, LatticeData lattice, double timescale);
+    void run_calculate_collision(dim3 blocks, dim3 threads, LatticeData lattice, float timescale);
 
     void run_calculate_reflections(dim3 blocks, dim3 threads, LatticeData lattice, LatticeData templattice);
